Move SettingsPanel components into own subfolder, and migrate to new PrimeVue Tabs component
