+++ src/components/SettingsPanel/ClickCountEditor.vue
